Move all test_struct.py tests
