+++ python/cudf/cudf/tests/dataframe/indexing/test_getitem.py
@@ -1,0 +1,10 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+
+import cudf
+from cudf.testing import assert_eq
+
+
+def test_struct_of_struct_loc():
+    df = cudf.DataFrame({"col": [{"a": {"b": 1}}]})
+    expect = cudf.Series([{"a": {"b": 1}}], name="col")
+    assert_eq(expect, df["col"])
