+++ python/cudf/cudf/tests/series/accessors/test_struct.py
@@ -56,3 +56,13 @@ def test_struct_explode():
     # check that a copy was made:
     got["a"][0] = 5
     assert_eq(s.struct.explode(), expect)
+
+
+@pytest.mark.parametrize(
+    "key, expect", [(0, [1, 3]), (1, [2, 4]), ("a", [1, 3]), ("b", [2, 4])]
+)
+def test_struct_for_field(key, expect):
+    sr = cudf.Series([{"a": 1, "b": 2}, {"a": 3, "b": 4}])
+    expect = cudf.Series(expect)
+    got = sr.struct.field(key)
+    assert_eq(expect, got)
